+++ frontend/compose-html-ext/src/jsMain/kotlin/com/varabyte/kobweb/compose/css/Table.kt
@@ -20,3 +20,15 @@ sealed interface CaptionSide : StylePropertyValue {
 fun StyleScope.captionSide(captionSide: CaptionSide) {
     property("caption-side", captionSide)
 }
+
+// https://developer.mozilla.org/en-US/docs/Web/CSS/empty-cells
+sealed interface EmptyCells : StylePropertyValue {
+    companion object : CssGlobalValues<EmptyCells> {
+        val Show get() = "show".unsafeCast<EmptyCells>()
+        val Hide get() = "hide".unsafeCast<EmptyCells>()
+    }
+}
+
+fun StyleScope.emptyCells(emptyCells: EmptyCells) {
+    property("empty-cells", emptyCells)
+}
